lots more restructuring
