+++ wicket/src/state/rack.rs
@@ -1,0 +1,122 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
+//! Global state of the rack.
+//!
+//! This is mostly visual selection state right now.
+
+use super::inventory::ComponentId;
+use slog::Logger;
+
+// Easter egg alert: Support for Knight Rider mode
+#[derive(Debug, Default)]
+pub struct KnightRiderMode {
+    pub count: usize,
+}
+
+impl KnightRiderMode {
+    pub fn step(&mut self) {
+        self.count += 1;
+    }
+}
+
+// The visual state of the rack
+#[derive(Debug)]
+pub struct RackState {
+    pub log: Option<Logger>,
+    pub selected: ComponentId,
+    pub knight_rider_mode: Option<KnightRiderMode>,
+
+    // Useful for arrow based navigation. When we cross the switches going up
+    // or down the rack we want to stay in the same column. This allows a user
+    // to repeatedly hit up or down arrow and stay in the same column, as they
+    // would expect.
+    //
+    // This is the only part of the state that is "historical", and doesn't
+    // rely on the `self.selected` at all times.
+    left_column: bool,
+}
+
+#[allow(clippy::new_without_default)]
+impl RackState {
+    pub fn new() -> RackState {
+        RackState {
+            log: None,
+            selected: ComponentId::Sled(0),
+            knight_rider_mode: None,
+
+            // Default to the left column, where sled 0 lives
+            left_column: true,
+        }
+    }
+
+    pub fn toggle_knight_rider_mode(&mut self) {
+        if self.knight_rider_mode.is_none() {
+            self.knight_rider_mode = Some(KnightRiderMode::default());
+        } else {
+            self.knight_rider_mode = None;
+        }
+    }
+
+    pub fn up(&mut self) {
+        self.selected = match self.selected {
+            ComponentId::Sled(i) if i == 14 || i == 15 => {
+                ComponentId::Switch(0)
+            }
+            ComponentId::Sled(i) => ComponentId::Sled((i + 2) % 32),
+            ComponentId::Switch(0) => ComponentId::Psc(0),
+            ComponentId::Switch(1) => {
+                if self.left_column {
+                    ComponentId::Sled(16)
+                } else {
+                    ComponentId::Sled(17)
+                }
+            }
+            ComponentId::Switch(_) => unreachable!(),
+            ComponentId::Psc(0) => ComponentId::Psc(1),
+            ComponentId::Psc(1) => ComponentId::Switch(1),
+            ComponentId::Psc(_) => unreachable!(),
+        };
+    }
+
+    pub fn down(&mut self) {
+        self.selected = match self.selected {
+            ComponentId::Sled(i) if i == 16 || i == 17 => {
+                ComponentId::Switch(1)
+            }
+            ComponentId::Sled(i) => ComponentId::Sled((30 + i) % 32),
+            ComponentId::Switch(1) => ComponentId::Psc(1),
+            ComponentId::Switch(0) => {
+                if self.left_column {
+                    ComponentId::Sled(14)
+                } else {
+                    ComponentId::Sled(15)
+                }
+            }
+            ComponentId::Switch(_) => unreachable!(),
+            ComponentId::Psc(0) => ComponentId::Switch(0),
+            ComponentId::Psc(1) => ComponentId::Psc(0),
+            ComponentId::Psc(_) => unreachable!(),
+        };
+    }
+
+    pub fn left_or_right(&mut self) {
+        match self.selected {
+            ComponentId::Sled(i) => {
+                if self.left_column {
+                    self.left_column = false;
+                    self.selected = ComponentId::Sled(i + 1);
+                } else {
+                    self.left_column = true;
+                    self.selected = ComponentId::Sled(i - 1);
+                }
+            }
+            _ => (),
+        }
+    }
+
+    pub fn set_logger(&mut self, log: Logger) {
+        self.log = Some(log);
+    }
+}
