+++ wicket/src/ui/widgets/animated_logo.rs
