+++ wicket/src/ui/widgets/banner.rs
